Signal the ascending date order (#600)
Maybe I'm tired, though I checked this page three time until I understood that the newest version row is in the end of the table. I'd suggest to flip the sort order, but I guess there is a reason to have the oldest first, so it might be a good idea to signal this order in the paragraph before.